pyproject.toml
[tool.poetry]
name = "leeds-hmo-pipeline"
version = "0.1.0"
description = ""
authors = ["christophercarlon <cmcarleone@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
duckdb = "0.9.2"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.1"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.2"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
